requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
flask-mail
pyjwt
flask-bootstrap
flask-moment
flask-babel
guess-language_spirit
requests
python-dotenv
elasticsearch
gunicorn
pymysql
psycopg2
rq